Move constraints to dbt.common
