+++ core/dbt/adapters/base/impl.py
@@ -23,7 +23,7 @@ from typing import (
 from multiprocessing.context import SpawnContext
 
 from dbt.adapters.capability import Capability, CapabilityDict
-from dbt.contracts.graph.nodes import ColumnLevelConstraint, ConstraintType, ModelLevelConstraint
+from dbt.common.constraints import ColumnLevelConstraint, ConstraintType, ModelLevelConstraint
 
 import agate
 import pytz
