+++ core/dbt/contracts/graph/nodes.py
@@ -2,12 +2,12 @@ import os
 from datetime import datetime
 import time
 from dataclasses import dataclass, field
-from enum import Enum
 import hashlib
 
 from mashumaro.types import SerializableType
 from typing import Optional, Union, List, Dict, Any, Sequence, Tuple, Iterator, Literal
 
+from dbt.common.constraints import ColumnLevelConstraint, ConstraintType, ModelLevelConstraint
 from dbt.common.dataclass_schema import dbtClassMixin, ExtensibleDbtClassMixin
 
 from dbt.clients.system import write_file
@@ -176,44 +176,6 @@ class RefArgs(dbtClassMixin):
             return {"version": self.version}
         else:
             return {}
-
-
-class ConstraintType(str, Enum):
-    check = "check"
-    not_null = "not_null"
-    unique = "unique"
-    primary_key = "primary_key"
-    foreign_key = "foreign_key"
-    custom = "custom"
-
-    @classmethod
-    def is_valid(cls, item):
-        try:
-            cls(item)
-        except ValueError:
-            return False
-        return True
-
-
-@dataclass
-class ColumnLevelConstraint(dbtClassMixin):
-    type: ConstraintType
-    name: Optional[str] = None
-    # expression is a user-provided field that will depend on the constraint type.
-    # It could be a predicate (check type), or a sequence sql keywords (e.g. unique type),
-    # so the vague naming of 'expression' is intended to capture this range.
-    expression: Optional[str] = None
-    warn_unenforced: bool = (
-        True  # Warn if constraint cannot be enforced by platform but will be in DDL
-    )
-    warn_unsupported: bool = (
-        True  # Warn if constraint is not supported by the platform and won't be in DDL
-    )
-
-
-@dataclass
-class ModelLevelConstraint(ColumnLevelConstraint):
-    columns: List[str] = field(default_factory=list)
 
 
 @dataclass
